backup/pre_hexagonal_20260128_115357/requirements.txt
# Asistente Andrea - Requirements
# Updated: 2026-01-07 - Relaxed strict pins to compatible versions (~=)

# Web Framework - UPDATED
fastapi~=0.128.0  # Allows 0.128.x updates
uvicorn[standard]~=0.40.0
websockets~=15.0.1
itsdangerous>=2.0.0  # Required for SessionMiddleware

# Database - UPDATED
sqlalchemy~=2.0.45
asyncpg~=0.31.0
alembic~=1.17.2

# AI/ML Services - UPDATED
azure-cognitiveservices-speech>=1.34.0
groq>=1.0.0
onnxruntime>=1.16.0  # Required for Silero VAD

# HTTP Client - UPDATED
httpx~=0.28.1

# Configuration - UPDATED
python-dotenv~=1.2.1
pydantic-settings~=2.12.0

# Forms - UPDATED
python-multipart~=0.0.21

# Templates - UPDATED
jinja2~=3.1.6

# Twilio - REMOVED: Not used in production (only Telnyx)
# twilio~=10.4.0.2

# Testing - UPDATED
pytest~=9.0.2
pytest-cov~=7.0.0
pytest-mock~=3.15.1
# pytest-asyncio - Commented: incompatible with Python 3.13, use Python 3.11

# Code Quality - NEW (Punto A2)
ruff~=0.8.6

# Rate Limiting - NEW (Punto A3)
slowapi~=0.1.9

# Webhook Security - UPDATED for CVE fixes (2026-01-07 Security Audit)
cryptography>=42.0.0,<45.0.0  # Docker-compatible (v44+ fails on ARM emulation), still CVE-free
h11>=0.16.0  # Fixed CVE-2025-43859
urllib3>=2.6.3  # Fixed CVE-2026-21441
httpcore>=1.0.9  # Updated for h11 compatibility

# Input Sanitization - NEW (Punto A5)
markupsafe~=2.1.5

# State Management - NEW (Punto A9)
redis[hiredis]~=5.0.1

# OpenAI (Azure) - NEW (Punto A1)
openai>=1.0.0

# audioop Replacement (Punto A11 - Python 3.13+ Compatibility)
numpy>=1.26.0  # Audio processing (replaces audioop)

# Monitoring (Punto B2)
prometheus-client>=0.20.0  # Metrics collection and export

# Distributed Tracing (Punto B3 - Optional)
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-httpx>=0.43b0
opentelemetry-exporter-otlp>=1.22.0

# Security Scanning
pip-audit~=2.10.0

# Logging & Tracing - NEW (Punto B1)
python-json-logger~=2.0.7
asgi-correlation-id~=4.3.1

# Additional for SQLite testing (optional, not for production)
aiosqlite~=0.22.1

# UPDATE HISTORY:
# 2026-01-07: Relaxed versions to ~= (Compatible Release)
# 2026-01-06: Major dependency update - all packages updated to latest versions
#   - SQLAl chemy: 2.0.25 → 2.0.45 (11 security patches)
#   - FastAPI: 0.109.0 → 0.128.0 (security + features)
#   - Uvicorn: 0.27.0 → 0.40.0 (performance)
#   - Azure SDK: 1.34.0 → 1.47.0 (13 version updates)
#   - pytest: 8.0.0 → 9.0.2 (latest)
#   - pytest-cov: 4.1.0 → 7.0.0 (major update)

# SECURITY NOTES:
# 1. Run pip-audit regularly: python -m pip_audit
# 2. Dependabot configured for weekly updates (.github/dependabot.yml)
# 3. All known vulnerabilities resolved as of 2026-01-06
# 4. Use Python 3.11 in production for maximum compatibility

# DEPLOYMENT NOTES:
# 1. asyncpg may fail on Windows without C compiler - use pre-built wheel or Linux
# 2. For Python 3.13+: audioop module removed, some tests will skip automatically
# 3. For Python 3.11/3.12: all tests functional
# 4. Always use virtual environment: python -m venv venv
# 5. Install with: pip install -r requirements.txt
# 6. Run migrations: alembic upgrade head
